+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - PR #631: Add nvcategory based ordinal label encoder
 - PR #670 Add test skipping functionality to build.sh
 - PR #681 Add wrappers MBSGDClassifier and MBSGDRegressor around SGD
+- PR #673: prims: reduce cols by key primitive
 
 ## Improvements
 
@@ -29,6 +30,7 @@
 - PR #662: C++: Bulk clang-format updates
 - PR #671: Added pickle pytests and correct pickling of Base class
 - PR #677: build: 'deep-clean' to build.sh to clean faiss build as well
+- PR #683: Use stateless c++ API in KNN so that it can be pickled properly
 
 ## Bug Fixes
 - PR #584: Added missing virtual destructor to deviceAllocator and hostAllocator